% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdc-anonymisation.R
\name{addnoise.numvar}
\alias{addnoise.numvar}
\title{adding noise to all numerical variables.}
\usage{
addnoise.numvar(demg, conf)
}
\description{
The numerical variables includes datetime which will be
treated as numeric by converting them to numeric. It will be converted back
to datetime format at them end.
}

