% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdc-anonymisation.R
\name{anony.var}
\alias{anony.var}
\title{Parse YAML configuration file}
\usage{
anony.var(conf)
}
\arguments{
\item{conf}{either the path of YAML configuration of a list}
}
\value{
a list contrains all the variable names broken down in each
category.
}
\description{
Parse the YAML configuration file and check the variable names.
}

