% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdc-anonymisation.R
\name{anonymisation}
\alias{anonymisation}
\title{Anonymise the critical care dataset}
\usage{
anonymisation(ccd, conf, remove.alive = T, verbose = F, ...)
}
\arguments{
\item{ccd}{identifiable data set in ccRecord format (see. ccdata R package)}

\item{conf}{yaml configuration which either characters of the path of yaml
file or the parsed list.}

\item{remove.alive}{logical value determines whether removing all alive
episode.}
}
\value{
ccRecord
}
\description{
This function is the sf
}

