% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.config.R
\name{create.sdc}
\alias{create.sdc}
\title{read.config}
\usage{
create.sdc(data, path)
}
\arguments{
\item{data}{the input confidential data in data.frame or data.table}

\item{config}{yaml file location}
}
\description{
read.config
}

