% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdc-anonymisation.R
\name{deltaTime1d}
\alias{deltaTime1d}
\title{Convert all the date-time columns to delta time, but still convert them back
with an origin of 1970-01-01.}
\usage{
deltaTime1d(demg, items, maxstay)
}
\description{
Convert all the date-time columns to delta time, but still convert them back
with an origin of 1970-01-01.
}

