% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.config.R
\name{do.sdc}
\alias{do.sdc}
\title{do.sdc}
\usage{
do.sdc(ccd, conf, remove.alive = T, verbose = F)
}
\arguments{
\item{ccd}{the identifiable ccRecord object}

\item{config}{yaml file location}
}
\description{
do.sdc
}

