% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdc-anonymisation.R
\name{microaggregation.numvar}
\alias{microaggregation.numvar}
\title{microaggregation on numeric variables}
\usage{
microaggregation.numvar(demg, conf)
}
\arguments{
\item{demg}{data.table/data.frame}

\item{conf}{configuration list}
}
\value{
demg data.frame
}
\description{
microaggregation on numeric variables
}

