% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-config.R
\name{parse.conf}
\alias{parse.conf}
\title{Parse YAML configuration file}
\usage{
parse.conf(conf)
}
\arguments{
\item{conf}{either the path of YAML configuration of a list}
}
\value{
a list contrains all the parse.conf broken down in each
category.
}
\description{
Parse the YAML configuration file and check the parse.conf.
}

