% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdc-anonymisation.R
\name{remove.direct.vars}
\alias{remove.direct.vars}
\title{Remove direct identifiers specified}
\usage{
remove.direct.vars(data, dirvs)
}
\arguments{
\item{data}{data.frame/data.table}

\item{dirvs}{direct variables name}
}
\value{
data data.frame/data/table
}
\description{
Remove direct identifiers specified
}

