% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdc-anonymisation.R
\name{sdc.trail}
\alias{sdc.trail}
\title{sdc.trail}
\usage{
sdc.trail(ccd, conf, remove.alive = T, verbose = F, k.anon = 20,
  l.div = 10)
}
\arguments{
\item{ccd}{the identifiable ccRecord object}

\item{config}{yaml file location}
}
\description{
sdc.trail
}

