% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdc-anonymisation.R
\name{sdc.trial}
\alias{sdc.trial}
\title{sdc.trial}
\usage{
sdc.trial(ccd, conf, remove.alive = T, verbose = F, k.anon = 20,
  l.div = NULL)
}
\arguments{
\item{ccd}{the identifiable ccRecord object}

\item{config}{yaml file location}
}
\description{
sdc.trial
}

