% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.conf.R
\name{template.conf}
\alias{template.conf}
\title{create a YAML configuration template}
\usage{
template.conf(outfile = "template.yaml")
}
\arguments{
\item{outfile}{the file name of YAML configuration template.}
}
\description{
This is a function to create a YAML configuration example which can be used
as a template for users.
}

