% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdc-anonymisation.R
\name{variables.name}
\alias{variables.name}
\title{Parse the confiuration file and check the variable names.}
\usage{
variables.name(conf)
}
\description{
Parse the confiuration file and check the variable names.
}

