% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.config.R
\name{create.sdc}
\alias{create.sdc}
\title{read.config}
\usage{
create.sdc(ccd, path, remove.alive = T)
}
\arguments{
\item{ccd}{the identifiable ccRecord object}

\item{config}{yaml file location}
}
\description{
read.config
}

